requirements.txt
MAMEToolkit>=1.1.0
numpy>=1.24.2
torch>=2.0.0
tensorboardX>=2.6
scipy>=1.10.1
pytorch-ignite